+++ app/components/layout/keyboard_layout/keyboard_layout.js
@@ -1,29 +1,24 @@
 // Copyright (c) 2017-present Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-import React, {Component} from 'react';
+import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
 import {KeyboardAvoidingView, Platform, View} from 'react-native';
 
-export default class KeyboardLayout extends Component {
+export default class KeyboardLayout extends PureComponent {
     static propTypes = {
         behaviour: PropTypes.string,
         children: PropTypes.node,
         keyboardVerticalOffset: PropTypes.number,
-        statusBarHeight: PropTypes.number,
-        watchStatusBar: PropTypes.bool
+        statusBarHeight: PropTypes.number
     };
 
     static defaultProps = {
         keyboardVerticalOffset: 0
     };
 
-    shouldComponentUpdate(nextProps) {
-        return nextProps.statusBarHeight !== this.props.statusBarHeight;
-    }
-
     render() {
-        const {behaviour, children, keyboardVerticalOffset, statusBarHeight, watchStatusBar, ...otherProps} = this.props;
+        const {behaviour, children, keyboardVerticalOffset, statusBarHeight, ...otherProps} = this.props;
 
         if (Platform.OS === 'android') {
             return (
@@ -34,7 +29,7 @@ export default class KeyboardLayout extends Component {
         }
 
         let height = 0;
-        if (watchStatusBar && statusBarHeight > 20) {
+        if (statusBarHeight > 20) {
             height = (statusBarHeight - 20) + keyboardVerticalOffset;
         } else {
             height = keyboardVerticalOffset;
